Optimizations to value Reduce
