+++ theorems/__init__.py
@@ -1,0 +1,12 @@
+from valuePercolate import valuePercolate
+from anglePercolate import anglePercolate
+from angleSponger import angleSponger
+from valueSponger import valueSponger
+from modifierSponger import modifierSponger
+from modifierReduce import modifierReduce
+from loopReduce import loopReduce
+from negativeReduce import negativeReduce
+from pushReduce import pushReduce
+from valueReduce import valueReduce
+
+from basics import *
