+++ theorems/angleSponger.py
@@ -1,0 +1,7 @@
+from basics import *
+import re
+
+def angleSponger(snippet):
+	snippet = snippet.replace("CC","")
+	snippet = cleanup(snippet)
+	return snippet
