+++ theorems/basics.py
@@ -42,6 +42,11 @@ def findMatch(snippet, index):
 
 def balanced(snippet):
 	stack = []
+	#Remove all non brace characters
+	while re.search("[^\(\)\{\}<>\[\]]",snippet):
+		snippet = snippet[:re.search("[^\(\)\{\}<>\[\]]",snippet).span()[0]]+snippet[re.search("[^\(\)\{\}<>\[\]]",snippet).span()[1]:]
+	if snippet == "":
+		return True
 	if snippet[0] in ")}>]":
 		return False
 	for character in snippet:
