+++ theorems/modifierReduce.py
@@ -1,0 +1,7 @@
+from basics import *
+import re
+
+def modifierReduce(snippet):
+	snippet = snippet.replace("><","")
+	snippet = snippet.replace("][","")
+	return snippet
